# pairhmm: small bitmask optimisation