The previous column bitmasks can be read again from the same array
instead of needing to remember them.

```
forward_f32x8           time:   [32.690 ms 32.707 ms 32.725 ms]
                        change: [-3.6103% -3.1147% -2.6176%] (p = 0.00 < 0.05)
                        Performance has improved.

forward_f32x16          time:   [16.230 ms 16.315 ms 16.440 ms]
                        change: [-3.9896% -3.3519% -2.5230%] (p = 0.00 < 0.05)
                        Performance has improved.

forward_f64x4           time:   [49.923 ms 50.104 ms 50.414 ms]
                        change: [+0.0725% +0.7525% +1.5384%] (p = 0.04 < 0.05)
                        Change within noise threshold.

forward_f64x8           time:   [25.710 ms 25.720 ms 25.733 ms]
                        change: [-2.4106% -2.1354% -1.8706%] (p = 0.00 < 0.05)
                        Performance has improved.
```